moved double function to cpt class
