+++ src/Wordpress/CustomPostType/Item.php
@@ -4,7 +4,6 @@ namespace CommonsBooking\Wordpress\CustomPostType;
 
 use CommonsBooking\Repository\UserRepository;
 use CommonsBooking\Settings\Settings;
-use CommonsBooking\View\Admin\Filter;
 
 class Item extends CustomPostType {
 
@@ -30,25 +29,7 @@ class Item extends CustomPostType {
 		add_action( 'pre_get_posts', array( $this, 'filterAdminList' ) );
 	}
 
-	/**
-	 * Adds Category filter to backend list view
-	 * 
-	 */
-	public static function addAdminCategoryFilter() {
-		$values = [];
-		$terms = get_terms(array(
-			'taxonomy'	=> self::$postType . 's_category'
-		));
-		foreach ( $terms as $term ) {
-			$values[ $term->term_id ] = $term->name;
-		}
-		Filter::renderFilter(
-			static::$postType,
-			esc_html__( 'Filter By Category ', 'commonsbooking' ),
-			'filter_post_category',
-			$values
-		);
-	}
+
 
 	/**
 	 * Filters admin list by type (e.g. bookable, repair etc. )
